requirements.txt
autopep8==1.4.3
certifi==2019.6.16
chardet==3.0.4
Django==2.1.7
django-crispy-forms==1.7.2
django-filter==2.1.0
djangorestframework==3.9.2
idna==2.8
Markdown==3.0.1
Pillow==5.4.1
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2018.9
pyuploadcare==2.6.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
